more safe timer calculator

+++ Homework/Item11/timer.h
@@ -24,11 +24,11 @@ enum // TMOD
 #define HIBYTE(x) ((unsigned char)((x)>>8))
 
 // 定时器定时计算 单位：us
-#define T8BH(x)  HIBYTE(0x100-(x)/(12/FREQ))
-#define T8BL(x)  LOBYTE(0x100-(x)/(12/FREQ))
-#define T13BH(x) HIBYTE(0x2000-(x)/(12/FREQ))
-#define T13BL(x) LOBYTE(0x2000-(x)/(12/FREQ))
-#define T16BH(x) HIBYTE(0x10000-(x)/(12/FREQ))
-#define T16BL(x) LOBYTE(0x10000-(x)/(12/FREQ))
+#define T8BH(x)  HIBYTE((unsigned short)(0x100-(x)/(12/FREQ)))
+#define T8BL(x)  LOBYTE((unsigned short)(0x100-(x)/(12/FREQ)))
+#define T13BH(x) HIBYTE((unsigned short)(0x2000-(x)/(12/FREQ)))
+#define T13BL(x) LOBYTE((unsigned short)(0x2000-(x)/(12/FREQ)))
+#define T16BH(x) HIBYTE((unsigned short)(0x10000-(x)/(12/FREQ)))
+#define T16BL(x) LOBYTE((unsigned short)(0x10000-(x)/(12/FREQ)))
 
 #endif // TIMER_H_
